tests/wpt/web-platform-tests/2dcontext/text-styles/2d.text.font.parse.size.percentage.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.text.font.parse.size.percentage</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.font.parse.size.percentage</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" style="font-size: 144px" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

ctx.font = '50% serif';
_assertSame(ctx.font, '72px serif', "ctx.font", "'72px serif'");
canvas.setAttribute('style', 'font-size: 100px');
_assertSame(ctx.font, '72px serif', "ctx.font", "'72px serif'");


});
</script>
